bot_project/requirements.txt

pyrogram
pytgcalls
yt-dlp
ffmpeg-python
